+++ LifeLibrary/LifeLibrary/Screen/Splash/SplashViewController.swift
@@ -32,12 +32,12 @@ class SplashViewController: UIViewController {
             .take(until: maxExposeTime)
             .subscribe(onNext: { [weak self] accessToken in
                 // 최소노출시간 대기 후 로그인여부가 확인되면 화면 진입
-                accessToken == "" ? self?.presentLoginView() : self?.presentMainView()
+                accessToken == "" ? self?.presentOnboardingView() : self?.presentMainView()
             }, onCompleted: { [weak self] in
                 // 최대노출시간을 초과한 후에도 현재 화면이 노출 중이면 다음 화면으로 강제 이동
                 guard self?.isVisible == true else { return }
                 
-                self?.presentLoginView()
+                self?.presentOnboardingView()
             }).disposed(by: disposeBag)
     }
 }
@@ -49,8 +49,8 @@ extension SplashViewController {
         loginAccessTokenRx.accept(accessToken)
     }
     
-    func presentLoginView() {
-        let vc = LoginViewController()
+    func presentOnboardingView() {
+        let vc = OnboardingViewController()
         
         vc.modalPresentationStyle = .overFullScreen
