+++ LifeLibrary/LifeLibrary/Screen/User/Login/LoginViewController.swift
@@ -9,21 +9,7 @@ import UIKit
 
 class LoginViewController: UIViewController {
     
-    @IBOutlet weak var loginButton: UIButton!
-    @IBOutlet weak var signupButton: UIButton!
-    
     override func viewDidLoad() {
         super.viewDidLoad()
-    }
-}
-
-// Touch Event
-extension LoginViewController {
-    @IBAction func onTouchLoginButton(_ sender: Any) {
-        
-    }
-    
-    @IBAction func onTouchSignUpButton(_ sender: Any) {
-        
     }
 }
